moving common exojax/radis functions to radis/api
