+++ radis/__init__.py
@@ -97,7 +97,10 @@ Parameters
 
 "OLDEST_COMPATIBLE_VERSION": "0.9.1"
     str: forces to regenerate cache files that were created in a previous version
-    See also : :py:func:`~radis.io.cache_files.load_h5_cache_file`
+
+    See Also
+    --------
+    :py:func:`~radis.api.cache_files.load_h5_cache_file`
 
 
 "USE_CYTHON": True
@@ -209,8 +212,9 @@ __all__ = [
     "__version__",
 ]
 
-# prevent cyclic importants:
-from . import db, io, lbl, los, misc, phys, spectrum, tools
+# prevent cyclic imports:
+from . import api, db, io, lbl, los, misc, phys, spectrum, tools
+from .api import *  # input / output common with ExoJax
 from .db import *  # database of molecules
 from .io import *  # input / output
 from .lbl import *  # line-by-line module
