+++ radis/api/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""Common API for RADIS & `Exojax <https://github.com/HajimeKawahara/exojax>`__
+
+-------------------------------------------------------------------------------
+
+"""
