+++ radis/api/cache_files.py
@@ -7,9 +7,9 @@ Routine Listing
 ---------------
 
 
-- :func:`~radis.io.cache_files.check_cache_file`
-- :func:`~radis.io.cache_files.check_not_deprecated`
-- :func:`~radis.io.cache_files.save_to_hdf`
+- :func:`~radis.api.cache_files.check_cache_file`
+- :func:`~radis.api.cache_files.check_not_deprecated`
+- :func:`~radis.api.cache_files.save_to_hdf`
 
 See Also
 --------
@@ -38,8 +38,8 @@ import radis
 
 try:
     from .hdf5 import DataFileManager
-except ImportError:  # if file is ran as a module
-    from radis.io.hdf5 import DataFileManager
+except ImportError:
+    from radis.api.hdf5 import DataFileManager
 from radis.misc.basics import compare_dict, is_float
 from radis.misc.printer import printm, printr
 from radis.misc.warning import DeprecatedFileWarning, IrrelevantFileWarning
@@ -251,7 +251,7 @@ def check_cache_file(
     - if ``'force'``, raise an error if file doesnt exist.
 
     Then look if it is deprecated (we just look at the attributes, the file
-    is never fully read). Deprecation is done by :py:func:`~radis.io.cache_files.check_not_deprecated`
+    is never fully read). Deprecation is done by :py:func:`~radis.api.cache_files.check_not_deprecated`
     comparing the ``metadata=`` content.
 
     - if deprecated, deletes it to regenerate later unless 'force' was used
@@ -282,7 +282,7 @@ def check_cache_file(
     See Also
     --------
 
-    :py:func:`~radis.io.cache_files.check_not_deprecated`
+    :py:func:`~radis.api.cache_files.check_not_deprecated`
     """
 
     # Test existence of file:
