+++ radis/io/cdsd.py
@@ -23,8 +23,8 @@ from collections import OrderedDict
 from os.path import exists, getmtime
 
 import radis
-from radis.io.cache_files import load_h5_cache_file, save_to_hdf
-from radis.io.hdf5 import DataFileManager
+from radis.api.cache_files import load_h5_cache_file, save_to_hdf
+from radis.api.hdf5 import DataFileManager
 from radis.io.tools import (
     drop_object_format_columns,
     parse_hitran_file,
@@ -154,7 +154,7 @@ def cdsd2df(
         will be left untouched.
     load_wavenum_min, load_wavenum_max: float
         if not ``'None'``, only load the cached file if it contains data for
-        wavenumbers above/below the specified value. See :py:func`~radis.io.cache_files.load_h5_cache_file`.
+        wavenumbers above/below the specified value. See :py:func`~radis.api.cache_files.load_h5_cache_file`.
         Default ``'None'``.
     engine: 'pytables', 'vaex'
         format for Hdf5 cache file. Default `pytables`
