+++ radis/io/hitran.py
@@ -31,6 +31,9 @@ from os.path import exists, getmtime, join, split
 import pandas as pd
 from numpy import int64
 
+from radis.api.cache_files import load_h5_cache_file, save_to_hdf
+from radis.api.dbmanager import DatabaseManager
+from radis.api.hdf5 import DataFileManager
 from radis.db.classes import (  # get_molecule_identifier,
     HITRAN_CLASS1,
     HITRAN_CLASS2,
@@ -50,28 +53,11 @@ from radis.db.classes import (  # get_molecule_identifier,
     HITRAN_GROUP6,
     get_molecule,
 )
-
-from ..misc.warning import AccuracyWarning
-
-try:
-    from .cache_files import load_h5_cache_file, save_to_hdf
-    from .dbmanager import DatabaseManager
-    from .hdf5 import DataFileManager
-    from .tools import (
-        drop_object_format_columns,
-        parse_hitran_file,
-        replace_PQR_with_m101,
-    )
-except ImportError:
-    from radis.io.cache_files import load_h5_cache_file, save_to_hdf
-    from radis.io.dbmanager import DatabaseManager
-    from radis.io.hdf5 import DataFileManager
-    from radis.io.tools import (
-        drop_object_format_columns,
-        parse_hitran_file,
-        replace_PQR_with_m101,
-    )
-
+from radis.io.tools import (
+    drop_object_format_columns,
+    parse_hitran_file,
+    replace_PQR_with_m101,
+)
 
 # %% Parsing functions
 
@@ -159,7 +145,7 @@ def hit2df(
         will be left untouched.
     load_wavenum_min, load_wavenum_max: float
         if not ``'None'``, only load the cached file if it contains data for
-        wavenumbers above/below the specified value. See :py:func`~radis.io.cache_files.load_h5_cache_file`.
+        wavenumbers above/below the specified value. See :py:func`~radis.api.cache_files.load_h5_cache_file`.
         Default ``'None'``.
     engine: 'pytables', 'vaex'
         format for Hdf5 cache file. Default `pytables`
